xishuipang-flat/tts/requirements.txt
# MeloTTS 本身（从 GitHub 装，不在 PyPI）
# 安装命令：pip install git+https://github.com/myshell-ai/MeloTTS.git
# 下面这些是它的运行时依赖 + 我们自己的工具

# MongoDB
pymongo>=4.6

# 环境变量
python-dotenv>=1.0

# 进度条
tqdm>=4.66

# HTTP 本地服务（测试用）
# 只用 stdlib 的 http.server，不加额外依赖

# 注意：melo-tts 建议单独装：
#   pip install git+https://github.com/myshell-ai/MeloTTS.git
#   python -m unidic download
#
# 如果遇到 torch/torchaudio 版本冲突，M1 Mac 推荐：
#   pip install torch==2.3.1 torchaudio==2.3.1
